@@ -9,119 +9,114 @@
   "collective": {
     "url": "https://opencollective.com/docsify"
   },
+  "keywords": [
+    "client",
+    "creator",
+    "crs",
+    "doc",
+    "docs",
+    "documentation",
+    "generator",
+    "markdown"
+  ],
+  "engines": {
+    "node": ">=20.11.0"
+  },
   "type": "module",
-  "// The 'main' and 'unpkg' fields will remain as legacy for backwards compatibility for now. We will add deprectaion warnings to these outputs to give people time to see the warnings in their apps in a non-breaking way, and eventually we can remove the legacy stuff.": "",
-  "main": "lib/docsify.js",
-  "unpkg": "lib/docsify.min.js",
-  "// We're using the 'exports' field as an override of the 'main' field to provide the new ESM setup. Once we remove legacy 'main', we will remove the 'exports' field and have a simple ESM setup with only a 'main' field.": "",
+  "main": "dist/docsify.js",
   "exports": {
-    ".": "./src/Docsify.js",
+    ".": "./src/core/Docsify.js",
     "./*": "./*"
   },
   "files": [
-    "lib",
+    "dist",
     "themes"
   ],
-  "scripts": {
-    "build:cover": "node build/cover.js",
-    "build:css:min": "mkdirp lib/themes && npm run css -- -o lib/themes && node build/mincss.js",
-    "build:css": "mkdirp themes && npm run css -- -o themes",
-    "build:emoji": "node ./build/emoji.js",
-    "build:js": "cross-env NODE_ENV=production node build/build.js",
-    "build:test": "npm run build && npm test",
-    "build": "rimraf lib themes && run-s build:js build:css build:css:min build:cover build:emoji",
-    "css": "node build/css",
-    "dev": "run-p serve watch:*",
-    "docker:build:test": "npm run docker:cli -- build:test",
-    "docker:build": "docker build -f Dockerfile -t docsify-test:local .",
-    "docker:clean": "docker rmi docsify-test:local",
-    "docker:cli": "docker run --rm -it --ipc=host --mount type=bind,source=$(pwd)/test,target=/app/test docsify-test:local",
-    "docker:rebuild": "npm run docker:clean && npm run docker:build",
-    "docker:test:e2e": "npm run docker:cli -- test:e2e",
-    "docker:test:integration": "npm run docker:cli -- test:integration",
-    "docker:test:unit": "npm run docker:cli -- test:unit",
-    "docker:test": "npm run docker:cli -- test",
-    "prettier": "prettier . --write",
-    "lint:fix": "eslint . --fix",
-    "lint": "prettier . --check && eslint .",
-    "postinstall": "opencollective-postinstall",
-    "prepare": "npm run build",
-    "pub:next": "cross-env RELEASE_TAG=next sh build/release.sh",
-    "pub": "sh build/release.sh",
-    "serve": "node server",
-    "test:e2e": "playwright test",
-    "test:integration": "npm run jest -- --selectProjects integration",
-    "test:unit": "npm run jest -- --selectProjects unit",
-    "test": "npm run jest && run-s test:e2e",
-    "jest": "cross-env NODE_OPTIONS=--experimental-vm-modules jest",
-    "watch:css": "npm run css -- -o themes -w",
-    "watch:js": "node build/build.js"
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
-  },
   "lint-staged": {
     "*.js": "eslint --fix"
   },
   "dependencies": {
-    "marked": "^4.2.12",
-    "medium-zoom": "^1.0.8",
+    "dexie": "^4.0.8",
+    "medium-zoom": "^1.1.0",
     "opencollective-postinstall": "^2.0.2",
     "prismjs": "^1.29.0",
-    "strip-indent": "^3.0.0",
-    "tinydate": "^1.3.0",
-    "tweezer.js": "^1.4.0"
+    "tinydate": "^1.3.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.11.6",
-    "@babel/eslint-parser": "^7.16.5",
+    "@babel/eslint-parser": "^7.24.5",
     "@babel/preset-env": "^7.11.5",
-    "@eslint/js": "^8.43.0",
-    "@playwright/test": "^1.18.1",
-    "@rollup/plugin-commonjs": "^25.0.2",
+    "@eslint/js": "^10.0.0",
+    "@playwright/test": "^1.44.0",
+    "@rollup/plugin-babel": "^6.0.4",
+    "@rollup/plugin-commonjs": "^26.0.1",
     "@rollup/plugin-node-resolve": "^15.1.0",
     "@rollup/plugin-replace": "^5.0.2",
     "@rollup/plugin-terser": "^0.4.3",
     "@types/eslint": "^8.40.2",
-    "autoprefixer-stylus": "^1.0.0",
     "axios": "^1.5.0",
-    "babel-jest": "^27.4.6",
-    "browser-sync": "^2.26.12",
-    "chokidar": "^3.4.2",
+    "browser-sync": "^3.0.2",
     "common-tags": "^1.8.0",
     "conventional-changelog-cli": "^3.0.0",
-    "copy-dir": "^1.2.0",
     "cross-env": "^7.0.3",
-    "cssnano": "^4.1.10",
-    "eslint": "^8.43.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-jest": "^27.2.2",
-    "eslint-plugin-playwright": "^0.16.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "globals": "^13.20.0",
-    "husky": "^8.0.3",
-    "jest": "^27.4.7",
-    "lint-staged": "^13.2.2",
-    "live-server": "^1.2.1",
-    "mkdirp": "^3.0.0",
+    "cssnano": "^7.0.1",
+    "eslint": "^9.3.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-jest": "^28.5.0",
+    "eslint-plugin-playwright": "^1.6.1",
+    "eslint-plugin-prettier": "^5.1.3",
+    "glob": "^10.3.15",
+    "globals": "^15.3.0",
+    "husky": "^9.0.11",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "lint-staged": "^15.2.2",
+    "marked": "^14.1.0",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.8.8",
-    "rimraf": "^3.0.0",
-    "rollup": "^3.25.3",
-    "serve-handler": "^6.1.2",
-    "stylus": "^0.60.0",
-    "vue2": "npm:vue@^2.6.12",
-    "vue3": "npm:vue@^3.0.0",
+    "postcss-cli": "^11.0.0",
+    "postcss-import": "^16.1.0",
+    "postcss-nesting": "^13.0.0",
+    "prettier": "^3.2.5",
+    "rimraf": "^5.0.7",
+    "rollup": "^4.17.2",
+    "rollup-plugin-import-css": "^3.5.0",
+    "vue": "^3.4.27",
     "xhr-mock": "^2.5.1"
   },
-  "keywords": [
-    "doc",
-    "docs",
-    "documentation",
-    "creator",
-    "generator"
-  ]
+  "scripts": {
+    "build:cover": "node build/cover.js",
+    "build:css": "postcss \"src/themes/*.css\" \"src/themes/**/[!_]*.css\" --base src/themes --dir dist/themes --map",
+    "build:css:min": "cross-env NODE_ENV='production' npm run build:css -- --ext .min.css",
+    "build:emoji": "node ./build/emoji.js",
+    "build:js": "rollup -c",
+    "build": "run-s clean build:js build:css build:css:min build:cover",
+    "clean": "rimraf --glob dist/** themes/** _playwright*/**",
+    "dev": "run-p serve:dev watch:*",
+    "docker:build:test": "npm run docker:cli -- build:test",
+    "docker:build": "docker build -f Dockerfile -t docsify-test:local .",
+    "docker:clean": "docker rmi docsify-test:local",
+    "docker:cli": "docker run --rm -it --ipc=host --mount type=bind,source=$(pwd)/test,target=/app/test docsify-test:local",
+    "docker:rebuild": "run-s docker:clean docker:build",
+    "docker:test:e2e": "npm run docker:cli -- test:e2e",
+    "docker:test:integration": "npm run docker:cli -- test:integration",
+    "docker:test:unit": "npm run docker:cli -- test:unit",
+    "docker:test": "npm run docker:cli -- test",
+    "lint:fix": "prettier . --write && eslint . --fix",
+    "lint": "prettier . --check && eslint .",
+    "postinstall": "opencollective-postinstall && npx husky install",
+    "prepare": "npm run build",
+    "prettier": "prettier . --write",
+    "pub:next": "cross-env RELEASE_TAG=next sh build/release.sh",
+    "pub": "sh build/release.sh",
+    "serve:dev": "npm run serve -- --dev",
+    "serve": "node server",
+    "test:e2e": "playwright test",
+    "test:e2e:chromium": "playwright test --project='chromium'",
+    "test:e2e:ui": "playwright test --ui",
+    "test:integration": "npm run test:jest -- --selectProjects integration",
+    "test:jest": "cross-env NODE_OPTIONS=--experimental-vm-modules jest",
+    "test:unit": "npm run test:jest -- --selectProjects unit",
+    "test": "run-s test:jest test:e2e",
+    "watch:css": "run-p 'build:css -- --watch' 'build:css:min -- --watch'",
+    "watch:js": "npm run build:js -- --watch"
+  }
 }
