@@ -9,7 +9,7 @@ import { get } from './util/ajax.js';
 // TODO This is deprecated, kept for backwards compatibility. Remove in a
 // major release. We'll tell people to get everything from the DOCSIFY global
 // when using the global build, but we'll highly recommend for them to import
-// from the ESM build (f.e. lib/docsify.esm.js and lib/docsify.min.esm.js).
+// from the ESM build (f.e. dist/docsify.esm.js and dist/docsify.min.esm.js).
 export default function initGlobalAPI() {
   window.Docsify = {
     util,
