@@ -5,62 +5,17 @@ import { tree as treeTpl } from './tpl.js';
 import { genTree } from './gen-tree.js';
 import { slugify } from './slugify.js';
 import { emojify } from './emojify.js';
-import {
-  getAndRemoveConfig,
-  removeAtag,
-  getAndRemoveDocisfyIgnoreConfig,
-} from './utils.js';
+import { getAndRemoveConfig } from './utils.js';
 import { imageCompiler } from './compiler/image.js';
+import { headingCompiler } from './compiler/heading.js';
 import { highlightCodeCompiler } from './compiler/code.js';
 import { paragraphCompiler } from './compiler/paragraph.js';
 import { taskListCompiler } from './compiler/taskList.js';
 import { taskListItemCompiler } from './compiler/taskListItem.js';
 import { linkCompiler } from './compiler/link.js';
+import { compileMedia } from './compiler/media.js';
 
 const cachedLinks = {};
-
-const compileMedia = {
-  markdown(url) {
-    return {
-      url,
-    };
-  },
-  mermaid(url) {
-    return {
-      url,
-    };
-  },
-  iframe(url, title) {
-    return {
-      html: `<iframe src="${url}" ${
-        title || 'width=100% height=400'
-      }></iframe>`,
-    };
-  },
-  video(url, title) {
-    return {
-      html: `<video src="${url}" ${title || 'controls'}>Not Support</video>`,
-    };
-  },
-  audio(url, title) {
-    return {
-      html: `<audio src="${url}" ${title || 'controls'}>Not Support</audio>`,
-    };
-  },
-  code(url, title) {
-    let lang = url.match(/\.(\w+)$/);
-
-    lang = title || (lang && lang[1]);
-    if (lang === 'md') {
-      lang = 'markdown';
-    }
-
-    return {
-      url,
-      lang,
-    };
-  },
-};
 
 export class Compiler {
   constructor(config, router) {
@@ -74,18 +29,17 @@ export class Compiler {
       this.linkTarget === '_blank' ? config.externalLinkRel || 'noopener' : '';
     this.contentBase = router.getBasePath();
 
-    const renderer = this._initRenderer();
-    this.heading = renderer.heading;
+    this.renderer = this._initRenderer();
     let compile;
     const mdConf = config.markdown || {};
 
     if (isFn(mdConf)) {
-      compile = mdConf(marked, renderer);
+      compile = mdConf(marked, this.renderer);
     } else {
       marked.setOptions(
         Object.assign(mdConf, {
-          renderer: Object.assign(renderer, mdConf.renderer),
-        })
+          renderer: Object.assign(this.renderer, mdConf.renderer),
+        }),
       );
       compile = marked;
     }
@@ -152,7 +106,7 @@ export class Compiler {
         href = getPath(
           this.contentBase,
           getParentPath(this.router.getCurrentPath()),
-          href
+          href,
         );
       }
 
@@ -174,7 +128,7 @@ export class Compiler {
           type = 'audio';
         }
 
-        embed = compileMedia[type].call(this, href, title);
+        embed = compileMedia[type](href, title);
         embed.type = type;
       }
 
@@ -199,42 +153,22 @@ export class Compiler {
   _initRenderer() {
     const renderer = new marked.Renderer();
     const { linkTarget, linkRel, router, contentBase } = this;
-    const _self = this;
+    // Supports mermaid
     const origin = {};
 
-    /**
-     * Render anchor tag
-     * @link https://github.com/markedjs/marked#overriding-renderer-methods
-     * @param {String} text Text content
-     * @param {Number} level Type of heading (h<level> tag)
-     * @returns {String} Heading element
-     */
-    origin.heading = renderer.heading = function (text, level) {
-      let { str, config } = getAndRemoveConfig(text);
-      const nextToc = { level, title: str };
-
-      const { content, ignoreAllSubs, ignoreSubHeading } =
-        getAndRemoveDocisfyIgnoreConfig(str);
-      str = content.trim();
-
-      nextToc.title = removeAtag(str);
-      nextToc.ignoreAllSubs = ignoreAllSubs;
-      nextToc.ignoreSubHeading = ignoreSubHeading;
-      const slug = slugify(config.id || str);
-      const url = router.toURL(router.getCurrentPath(), { id: slug });
-      nextToc.slug = url;
-      _self.toc.push(nextToc);
-
-      return `<h${level} id="${slug}"><a href="${url}" data-id="${slug}" class="anchor"><span>${str}</span></a></h${level}>`;
-    };
-
+    // renderer customizers
+    origin.heading = headingCompiler({
+      renderer,
+      router,
+      compiler: this,
+    });
     origin.code = highlightCodeCompiler({ renderer });
     origin.link = linkCompiler({
       renderer,
       router,
       linkTarget,
       linkRel,
-      compilerClass: _self,
+      compiler: this,
     });
     origin.paragraph = paragraphCompiler({ renderer });
     origin.image = imageCompiler({ renderer, contentBase, router });
@@ -247,8 +181,8 @@ export class Compiler {
   }
 
   /**
-   * Compile sidebar
-   * @param {String} text Text content
+   * Compile sidebar, it uses  _sidebar.md ( or specific file) or the content's headings toc to render sidebar.
+   * @param {String} text Text content from the sidebar file, maybe empty
    * @param {Number} level Type of heading (h<level> tag)
    * @returns {String} Sidebar element
    */
@@ -257,32 +191,39 @@ export class Compiler {
     const currentPath = this.router.getCurrentPath();
     let html = '';
 
+    // compile sidebar from _sidebar.md
     if (text) {
-      html = this.compile(text);
-    } else {
-      for (let i = 0; i < toc.length; i++) {
-        if (toc[i].ignoreSubHeading) {
-          const deletedHeaderLevel = toc[i].level;
-          toc.splice(i, 1);
-          // Remove headers who are under current header
-          for (
-            let j = i;
-            j < toc.length && deletedHeaderLevel < toc[j].level;
-            j++
-          ) {
-            toc.splice(j, 1) && j-- && i++;
-          }
-
-          i--;
+      return this.compile(text);
+    }
+    // compile sidebar from content's headings toc
+    for (let i = 0; i < toc.length; i++) {
+      if (toc[i].ignoreSubHeading) {
+        const deletedHeaderLevel = toc[i].depth;
+        toc.splice(i, 1);
+        // Remove headers who are under current header
+        for (
+          let j = i;
+          j < toc.length && deletedHeaderLevel < toc[j].depth;
+          j++
+        ) {
+          toc.splice(j, 1) && j-- && i++;
         }
-      }
 
-      const tree = this.cacheTree[currentPath] || genTree(toc, level);
-      html = treeTpl(tree, /* html */ `<ul>{inner}</ul>`);
-      this.cacheTree[currentPath] = tree;
+        i--;
+      }
     }
 
+    const tree = this.cacheTree[currentPath] || genTree(toc, level);
+    html = treeTpl(tree);
+    this.cacheTree[currentPath] = tree;
     return html;
+  }
+
+  /**
+   * When current content redirect to a new path file, clean pre content headings toc
+   */
+  resetToc() {
+    this.toc = [];
   }
 
   /**
@@ -291,16 +232,12 @@ export class Compiler {
    * @returns {String} Sub-sidebar element
    */
   subSidebar(level) {
-    if (!level) {
-      this.toc = [];
-      return;
-    }
-
     const currentPath = this.router.getCurrentPath();
     const { cacheTree, toc } = this;
 
     toc[0] && toc[0].ignoreAllSubs && toc.splice(0);
-    toc[0] && toc[0].level === 1 && toc.shift();
+    // remove the first heading from the toc if it is a top-level heading
+    toc[0] && toc[0].depth === 1 && toc.shift();
 
     for (let i = 0; i < toc.length; i++) {
       toc[i].ignoreSubHeading && toc.splice(i, 1) && i--;
@@ -313,12 +250,21 @@ export class Compiler {
     return treeTpl(tree);
   }
 
+  /**
+   * Compile the text to generate HTML heading element based on the level
+   * @param {*} text Text content, for now it is only from the _sidebar.md file
+   * @param {*} level Type of heading (h<level> tag), for now it is always 1
+   * @returns
+   */
   header(text, level) {
-    return this.heading(text, level);
-  }
-
-  article(text) {
-    return this.compile(text);
+    const tokenHeading = {
+      type: 'heading',
+      raw: text,
+      depth: level,
+      text: text,
+      tokens: [{ type: 'text', raw: text, text: text }],
+    };
+    return this.renderer.heading(tokenHeading);
   }
 
   /**
