@@ -2,9 +2,9 @@ import { getAndRemoveConfig } from '../utils.js';
 import { isAbsolutePath, getPath, getParentPath } from '../../router/util.js';
 
 export const imageCompiler = ({ renderer, contentBase, router }) =>
-  (renderer.image = (href, title, text) => {
+  (renderer.image = ({ href, title, text }) => {
     let url = href;
-    let attrs = [];
+    const attrs = [];
 
     const { str, config } = getAndRemoveConfig(title);
     title = str;
@@ -39,6 +39,6 @@ export const imageCompiler = ({ renderer, contentBase, router }) =>
     }
 
     return /* html */ `<img src="${url}" data-origin="${href}" alt="${text}" ${attrs.join(
-      ' '
+      ' ',
     )} />`;
   });
