@@ -6,28 +6,33 @@ export const linkCompiler = ({
   router,
   linkTarget,
   linkRel,
-  compilerClass,
+  compiler,
 }) =>
-  (renderer.link = (href, title = '', text) => {
-    let attrs = [];
+  (renderer.link = function ({ href, title = '', tokens }) {
+    const attrs = [];
+    const text = this.parser.parseInline(tokens) || '';
     const { str, config } = getAndRemoveConfig(title);
     linkTarget = config.target || linkTarget;
     linkRel =
       linkTarget === '_blank'
-        ? compilerClass.config.externalLinkRel || 'noopener'
+        ? compiler.config.externalLinkRel || 'noopener'
         : '';
     title = str;
 
     if (
       !isAbsolutePath(href) &&
-      !compilerClass._matchNotCompileLink(href) &&
+      !compiler._matchNotCompileLink(href) &&
       !config.ignore
     ) {
-      if (href === compilerClass.config.homepage) {
+      if (href === compiler.config.homepage) {
         href = 'README';
       }
 
       href = router.toURL(href, null, router.getCurrentPath());
+
+      if (config.target) {
+        href.indexOf('mailto:') !== 0 && attrs.push(`target="${linkTarget}"`);
+      }
     } else {
       if (!isAbsolutePath(href) && href.slice(0, 2) === './') {
         href =
@@ -38,8 +43,8 @@ export const linkCompiler = ({
         href.indexOf('mailto:') === 0
           ? ''
           : linkRel !== ''
-          ? ` rel="${linkRel}"`
-          : ''
+            ? ` rel="${linkRel}"`
+            : '',
       );
     }
 
