@@ -1,12 +1,14 @@
 import { helper as helperTpl } from '../tpl.js';
 
 export const paragraphCompiler = ({ renderer }) =>
-  (renderer.paragraph = text => {
+  (renderer.paragraph = function ({ tokens }) {
+    const text = this.parser.parseInline(tokens);
     let result;
-    if (/^!&gt;/.test(text)) {
-      result = helperTpl('tip', text);
-    } else if (/^\?&gt;/.test(text)) {
-      result = helperTpl('warn', text);
+
+    if (text.startsWith('!&gt;')) {
+      result = helperTpl('callout important', text);
+    } else if (text.startsWith('?&gt;')) {
+      result = helperTpl('callout tip', text);
     } else {
       result = /* html */ `<p>${text}</p>`;
     }
