@@ -1,7 +1,16 @@
 export const taskListCompiler = ({ renderer }) =>
-  (renderer.list = (body, ordered, start) => {
+  (renderer.list = function (token) {
+    const ordered = token.ordered;
+    const start = token.start;
+
+    let body = '';
+    for (let j = 0; j < token.items.length; j++) {
+      const item = token.items[j];
+      body += this.listitem?.(item);
+    }
+
     const isTaskList = /<li class="task-list-item">/.test(
-      body.split('class="task-list"')[0]
+      body.split('class="task-list"')[0],
     );
     const isStartReq = start && start > 1;
     const tag = ordered ? 'ol' : 'ul';
