@@ -31,17 +31,17 @@ export function emojify(text, useNativeEmoji) {
       // Mark colons in tags
       .replace(
         /<(code|pre|script|template)[^>]*?>[\s\S]+?<\/(code|pre|script|template)>/g,
-        m => m.replace(/:/g, '__colon__')
+        m => m.replace(/:/g, '__colon__'),
       )
       // Mark colons in comments
       .replace(/<!--[\s\S]+?-->/g, m => m.replace(/:/g, '__colon__'))
       // Mark colons in URIs
       .replace(/([a-z]{2,}:)?\/\/[^\s'">)]+/gi, m =>
-        m.replace(/:/g, '__colon__')
+        m.replace(/:/g, '__colon__'),
       )
       // Replace emoji shorthand codes
       .replace(/:([a-z0-9_\-+]+?):/g, (m, $1) =>
-        replaceEmojiShorthand(m, $1, useNativeEmoji)
+        replaceEmojiShorthand(m, $1, useNativeEmoji),
       )
       // Restore colons in tags and comments
       .replace(/__colon__/g, ':')
