@@ -1,5 +1,4 @@
 import { isFn } from '../util/core.js';
-import { inBrowser } from './env.js';
 
 const cacheNode = {};
 
@@ -21,11 +20,24 @@ export function getNode(el, noCache = false) {
   return el;
 }
 
-export const $ = inBrowser && document;
+/**
+ *
+ * @param {*} el the targt element or the selector
+ * @param {*} content the content to be rendered as HTML
+ * @param {*} replace To replace the content (true) or insert instead (false) , default is false
+ */
+export function setHTML(el, content, replace) {
+  const node = getNode(el);
+  if (node) {
+    node[replace ? 'outerHTML' : 'innerHTML'] = content;
+  }
+}
 
-export const body = inBrowser && $.body;
+export const $ = document;
 
-export const head = inBrowser && $.head;
+export const body = $.body;
+
+export const head = $.head;
 
 /**
  * Find elements
