@@ -1,3 +1,9 @@
-export const inBrowser = true; // True for now, may change when we add SSR.
+const computedStyle = getComputedStyle(document.documentElement, null);
 
-export const isMobile = inBrowser && document.body.clientWidth <= 600;
+export const mobileBreakpoint = computedStyle.getPropertyValue(
+  '--_mobile-breakpoint',
+);
+
+export function isMobile() {
+  return window?.matchMedia?.(`(max-width: ${mobileBreakpoint})`)?.matches;
+}
