@@ -47,7 +47,7 @@ var errors = [],
     item: new RegExp('^-\\s+'),
     trim: new RegExp('^\\s+|\\s+$'),
     comment: new RegExp(
-      '([^\\\'\\"#]+([\\\'\\"][^\\\'\\"]*[\\\'\\"])*)*(#.*)?'
+      '([^\\\'\\"#]+([\\\'\\"][^\\\'\\"]*[\\\'\\"])*)*(#.*)?',
     ),
   };
 
