@@ -1,4 +1,3 @@
-/* eslint-disable no-unused-vars */
 import {
   init as initComponent,
   update as updateComponent,
@@ -11,9 +10,11 @@ const CONFIG = {
   paths: 'auto',
   depth: 2,
   maxAge: 86400000, // 1 day
-  hideOtherSidebarContent: false,
   namespace: undefined,
   pathNamespaces: undefined,
+  keyBindings: ['/', 'meta+k', 'ctrl+k'],
+  insertAfter: undefined, // CSS selector
+  insertBefore: undefined, // CSS selector
 };
 
 const install = function (hook, vm) {
@@ -28,14 +29,33 @@ const install = function (hook, vm) {
     CONFIG.placeholder = opts.placeholder || CONFIG.placeholder;
     CONFIG.noData = opts.noData || CONFIG.noData;
     CONFIG.depth = opts.depth || CONFIG.depth;
-    CONFIG.hideOtherSidebarContent =
-      opts.hideOtherSidebarContent || CONFIG.hideOtherSidebarContent;
     CONFIG.namespace = opts.namespace || CONFIG.namespace;
     CONFIG.pathNamespaces = opts.pathNamespaces || CONFIG.pathNamespaces;
+    CONFIG.keyBindings = opts.keyBindings || CONFIG.keyBindings;
   }
 
   const isAuto = CONFIG.paths === 'auto';
 
+  hook.init(() => {
+    const { keyBindings } = vm.config;
+
+    // Add key bindings
+    if (keyBindings.constructor === Object) {
+      keyBindings.focusSearch = {
+        bindings: CONFIG.keyBindings,
+        callback(e) {
+          const sidebarElm = document.querySelector('.sidebar');
+          const sidebarToggleElm = document.querySelector('.sidebar-toggle');
+          const searchElm = sidebarElm?.querySelector('input[type="search"]');
+          const isSidebarHidden = sidebarElm?.getBoundingClientRect().x < 0;
+
+          isSidebarHidden && sidebarToggleElm?.click();
+
+          setTimeout(() => searchElm?.focus(), isSidebarHidden ? 250 : 0);
+        },
+      };
+    }
+  });
   hook.mounted(_ => {
     initComponent(CONFIG, vm);
     !isAuto && initSearch(CONFIG, vm);
