@@ -37,8 +37,8 @@ test.describe('Gtag Plugin Tests', () => {
       config: {
         gtag: gtagList[0],
       },
-      scriptURLs: ['/lib/plugins/gtag.min.js'],
-      styleURLs: ['/lib/themes/vue.css'],
+      scriptURLs: ['/dist/plugins/gtag.js'],
+      styleURLs: ['/dist/themes/core.css'],
     };
 
     await docsifyInit({
@@ -63,8 +63,8 @@ test.describe('Gtag Plugin Tests', () => {
       config: {
         gtag: gtagList,
       },
-      scriptURLs: ['/lib/plugins/gtag.min.js'],
-      styleURLs: ['/lib/themes/vue.css'],
+      scriptURLs: ['/dist/plugins/gtag.js'],
+      styleURLs: ['/dist/themes/core.css'],
     };
 
     await docsifyInit({
@@ -80,11 +80,5 @@ test.describe('Gtag Plugin Tests', () => {
 
     // Tests
     expect($docsify.gtag).not.toEqual('');
-  });
-
-  test('data-ga attribute', async ({ page }) => {
-    pageRequestListened(page);
-
-    // TODO
   });
 });
