@@ -4,15 +4,15 @@ import { test, expect } from './fixtures/docsify-init-fixture.js';
 test.describe('Index file hosting', () => {
   const sharedOptions = {
     config: {
-      basePath: '/docs/index.html#/',
+      basePath: '/index.html#/',
     },
-    testURL: '/docs/index.html#/',
+    testURL: '/index.html#/',
   };
 
   test('should serve from index file', async ({ page }) => {
     await docsifyInit(sharedOptions);
     await expect(page.locator('#main')).toContainText(
-      'A magical documentation site generator'
+      'A magical documentation site generator',
     );
     expect(page.url()).toMatch(/index\.html#\/$/);
   });
