@@ -13,7 +13,7 @@ describe('Docs Site', function () {
 
     await docsifyInit({
       config: {
-        coverpage: 'docs/_coverpage.md',
+        coverpage: '_coverpage.md',
       },
       markdown: {
         homepage: '# Hello World',
@@ -32,7 +32,7 @@ describe('Docs Site', function () {
   test('sidebar renders and is unchanged', async () => {
     await docsifyInit({
       config: {
-        loadSidebar: 'docs/_sidebar.md',
+        loadSidebar: '_sidebar.md',
       },
       markdown: {
         homepage: '# Hello World',
@@ -50,7 +50,7 @@ describe('Docs Site', function () {
   test('navbar renders and is unchanged', async () => {
     await docsifyInit({
       config: {
-        loadNavbar: 'docs/_navbar.md',
+        loadNavbar: '_navbar.md',
       },
       markdown: {
         homepage: '# Hello World',
